Split property component tree into its own directory
