+++ .eslintrc.js
@@ -29,6 +29,7 @@ module.exports = {
   settings: {
     "import/resolver": {
       "node": {
+        "paths": ['src'],
         "extensions": ['.js','.jsx']
       }
     },
