+++ src/components/editor/ResourceTemplateForm.jsx
@@ -4,9 +4,9 @@ import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import PropTypes from 'prop-types'
 import shortid from 'shortid'
-import PropertyPanel from './PropertyPanel'
-import PropertyResourceTemplate from './PropertyResourceTemplate'
-import PropertyComponent from './PropertyComponent'
+import PropertyPanel from './property/PropertyPanel'
+import PropertyResourceTemplate from './property/PropertyResourceTemplate'
+import PropertyComponent from './property/PropertyComponent'
 import { removeAllContent, setItems } from '../../actions/index'
 import { getResourceTemplate } from '../../sinopiaServer'
 import { isResourceWithValueTemplateRef, resourceToName } from '../../Utilities'
