+++ src/components/editor/property/InputListLOC.jsx
@@ -5,8 +5,8 @@ import { Typeahead } from 'react-bootstrap-typeahead'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import shortid from 'shortid'
-import { changeSelections } from '../../actions/index'
-import { booleanPropertyFromTemplate, defaultValuesFromPropertyTemplate } from '../../Utilities'
+import { changeSelections } from '../../../actions/index'
+import { booleanPropertyFromTemplate, defaultValuesFromPropertyTemplate } from '../../../Utilities'
 
 class InputListLOC extends Component {
   constructor(props) {
