+++ src/components/editor/property/PropertyTemplateOutline.jsx
@@ -5,8 +5,8 @@ import PropTypes from 'prop-types'
 import shortid from 'shortid'
 import OutlineHeader from './OutlineHeader'
 import PropertyTypeRow from './PropertyTypeRow'
-import { getResourceTemplate } from '../../sinopiaServer'
-import { isResourceWithValueTemplateRef, resourceToName, templateBoolean } from '../../Utilities'
+import { getResourceTemplate } from '../../../sinopiaServer'
+import { isResourceWithValueTemplateRef, resourceToName, templateBoolean } from '../../../Utilities'
 import PropertyComponent from './PropertyComponent'
 import ResourceProperty from './ResourceProperty'
